Render Newton fractal for (z^3 - 1).

+++ sources/color_effects.c
@@ -6,7 +6,7 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/15 16:44:29 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/02/20 02:17:47 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/02/20 16:52:38 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int	lerp_color(t_fractol *ctl, int iterations)
 
 	normalized_iterations = ft_map_d((t_map_d){iterations, 0,
 			ctl->max_iterations, 0, 1});
-	intensity = ft_map_d((t_map_d){sqrt(normalized_iterations), 0, 1, 0,
+	intensity = ft_map_d((t_map_d){sqrtl(normalized_iterations), 0, 1, 0,
 			255.999});
 	color = int_color_lerp_wsteps(ctl->lerp_from, ctl->lerp_to, ctl->lerp_steps,
 			intensity);
